requirements.txt
环境配置：
_libgcc_mutex=0.1
_openmp_mutex=4.5
absl-py=1.4.0
astor=0.8.1
astunparse=1.6.3
ca-certificates=2023.5.7
cachetools=4.2.4
certifi=2023.5.7
charset-normalizer=3.2.0
cudatoolkit=10.1.243
cudnn=7.6.0
cycler=0.11.0
flatbuffers	=23.5.26
fonttools=4.38.0
gast=0.2.2
google-auth=1.35.0
google-auth-oauthlib=0.4.6
google-pasta=0.2.0
grpcio=1.56.0
h5py=2.10.0
idna=3.4
importlib-metadata=6.7.0
keras=2.3.1
keras-applications=1.0.8
keras-preprocessing=1.1.2
kiwisolver=1.4.4
ld_impl_linux-64=2.40
libclang=16.0.0
libffi=3.4.2
libgcc-ng=13.1.0
libgomp=13.1.0
libnsl=2.0.0
libsqlite=3.42.0
libstdcxx-ng=13.1.0
libzlib=1.2.13
markdown=3.4.3
markupsafe=2.1.3
matplotlib	=3.5.3
ncurses=6.4
numpy=1.21.6
oauthlib=3.2.2
openssl=3.1.1
opt-einsum=3.3.0
packaging	=23.1
pillow=9.5.0
pip=23.1.2
protobuf=3.19.6
pyasn1=0.5.0
pyasn1-modules=0.3.0
pyparsing=3.1.0
python=3.7.12
python-dateutil=2.8.2
pyyaml=6.0
readline=8.2
requests=2.31.0
requests-oauthlib=1.3.1
rsa=4.9
scipy=1.4.1
setuptools	=68.0.0
six=1.16.0
sqlite=3.42.0
tensorboard=2.1.1
tensorboard-data-server=0.6.1
tensorboard-plugin-wit=1.8.1
tensorflow-estimator=2.1.0
tensorflow-gpu=2.1.0
tensorflow-io-gcs-filesystem=0.32.0	
termcolor	=2.3.0
tk=8.6.12
typing-extensions=4.7.1
urllib3=1.26.16
werkzeug=2.2.3
wheel=0.40.0
wrapt=1.15.0
xz=5.2.6
zipp=3.15.0